stock_news_pipeline/requirements.txt
yfinance
pandas
psycopg2-binary
apache-airflow
beautifulsoup4
requests
python-dotenv
